:recycle: [SharedDocument] Refacto entity
