+++ config/doctrine/User.orm.yml
@@ -212,7 +212,7 @@ App\Entity\User:
             mappedBy: user
             orphanRemoval: true
         sharedDocuments:
-            targetEntity: App\Entity\SharedDocument
+            targetEntity: App\Entity\Attributes\SharedDocument
             mappedBy: sharedBy
             cascade: [ "remove" ]
     manyToOne:
