+++ src/DataFixtures/FixturesPurger.php
@@ -5,6 +5,7 @@ namespace App\DataFixtures;
 use App\Entity\AccessToken;
 use App\Entity\Administrateur;
 use App\Entity\Association;
+use App\Entity\Attributes\SharedDocument;
 use App\Entity\Beneficiaire;
 use App\Entity\BeneficiaireCentre;
 use App\Entity\Centre;
@@ -21,7 +22,6 @@ use App\Entity\Membre;
 use App\Entity\MembreCentre;
 use App\Entity\Note;
 use App\Entity\RefreshToken;
-use App\Entity\SharedDocument;
 use App\Entity\TypeCentre;
 use App\Entity\User;
 use Doctrine\Common\DataFixtures\Purger\ORMPurgerInterface;
