+++ src/Entity/Attributes/SharedDocument.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace App\Entity\Attributes;
+
+use App\Entity\Document;
+use App\Repository\SharedDocumentRepository;
+use Doctrine\ORM\Mapping as ORM;
+
+#[ORM\Entity(repositoryClass: SharedDocumentRepository::class)]
+#[ORM\Table(name: 'shareddocument')]
+class SharedDocument extends SharedPersonalData
+{
+    #[ORM\ManyToOne(targetEntity: Document::class, inversedBy: 'sharedDocuments')]
+    private ?Document $document = null;
+
+    public function __construct()
+    {
+        $this->sharedAt = new \DateTime('now');
+    }
+
+    public function getDocument(): ?Document
+    {
+        return $this->document;
+    }
+
+    public function getDocumentKey(): ?string
+    {
+        return $this->document?->getObjectKey();
+    }
+
+    public function setDocument(Document $document): self
+    {
+        $this->document = $document;
+
+        return $this;
+    }
+}
