+++ src/Entity/User.php
@@ -10,6 +10,7 @@ use App\Api\Dto\UserDto;
 use App\Api\Filters\UsernameFilter;
 use App\Api\State\SearchBeneficiaryProvider;
 use App\Api\State\UserStateProcessor;
+use App\Entity\Attributes\SharedDocument;
 use App\Repository\UserRepository;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
