+++ src/Factory/SharedDocumentFactory.php
@@ -2,8 +2,8 @@
 
 namespace App\Factory;
 
+use App\Entity\Attributes\SharedDocument;
 use App\Entity\Document;
-use App\Entity\SharedDocument;
 use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
 use Lexik\Bundle\JWTAuthenticationBundle\Services\JWTTokenManagerInterface;
